pyproject.toml
[project]
name = "mlh-pe-hackathon"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faker>=33.0",
    "flask>=3.1",
    "peewee>=3.17",
    "psycopg2-binary>=2.9",
    "pytest>=9.0.2",
    "python-dotenv>=1.0",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]